% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Emisja.R
\docType{package}
\name{Emisja}
\alias{Emisja}
\title{Pakiet Emisja}
\description{
Pakiet oblicza emisje EMEP/EEA zanieczyszczenia powietrza wybranego rodzaju pojazdu,
a takze wizualizujacy za pomoca wykresu wartosci wynikowe.
}
\details{
Pakiet zawiera dwie funkcje:

\code{obl_emisja()} oblicza emisje
\code{obl_emisja()} tworzy wykres
}
\section{Dane do obliczen}{

Funkcje pakietu korzystaja z danych dolaczonych do pakietu.

Dane sa dostepne dla uzytkownika jako obiekty \code{input}, \code{wskazniki}.
}

