% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R, R/input.R
\docType{data}
\name{input}
\alias{input}
\title{Plik zawierajacy dane zewnetrzne konieczne do dzialania pakietu}
\format{
Opisy znajduja sie w dokumentacji

Ramka danych zawiera:
\describe{
\item{Nat}{double,Natezenie ruchu}
\item{Segment}{character,Segment pojazdu np. Small}
\item{Fuel}{character,Rodzaj paliwa np. diesel}
\item{Technology}{character,Technologia stosowana do filtrowania zanieczyszczen np. DPF}
}
}
\source{
Obiekt wskazniki zawiera dane pochodzace
z EEA - Europejskiej Agnecji Srodowiska.
Obiekt input zawiera losowe wartosci natezenia ruchu.
}
\usage{
input

input
}
\description{
Dane zawieraja natezenie ruchu wygenerowane losowo dla konkretnych kategorii.
}
\examples{
\dontrun{
 input
}
\dontrun{
input
}
}
\keyword{datasets}
