% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obl_emisja.R
\name{obl_emisja}
\alias{obl_emisja}
\title{Funkcja obliczenia emisji spalin}
\usage{
obl_emisja(
  dane = input,
  kategoria = "Passenger Cars",
  euro = "Euro 5",
  mode = "",
  substancja = c("CO", "EC")
)
}
\arguments{
\item{dane}{data_frame - dane wejsciowe}

\item{kategoria}{character - kategoria pojazdu}

\item{euro}{character - europejski standard emisji spalin}

\item{mode}{character - miejsce pomiaru}

\item{substancja}{character - typ emitowanego zanieczyszczenia}
}
\value{
data_frame
}
\description{
Funkcja oblicza emisje korzystajac z danych zawartych w obiekcie
\code{wskazniki}, a tazke \code{input}.
W obiekcie \code{wskazniki} zapisane sa dane o emisji pobrane z EEA.
W obiekcie \code{input} zapisane sa stworzone losowo dane o natezeniu ruchu.
}
\details{
Wzor na obliczenie emisji:
Emisja = Nat x ((Alpha x Procent ^ 2 + Beta x Procent + Gamma + (Delta/Procent))/
(Epsilon x Procent ^ 2 + Zita x Procent + Hta) x (1-Reduction))
}
