% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wykres_emisja.R
\name{wykres_emisja}
\alias{wykres_emisja}
\title{Funkcja do Wizualizacji danych dostarczonych z pakietem}
\usage{
wykres_emisja(
  data = emisja_wynik,
  one_kategoria = Category,
  one_wartosc = "Passenger Cars",
  two_kategoria = Segment,
  two_wartosc = Emisja
)
}
\arguments{
\item{data}{data.frame}

\item{one_kategoria}{char - Wybor pierwszej kolumny z data}

\item{one_wartosc}{char - wybor wartosci z pierwszej kolumny}

\item{two_kategoria}{char - wybor 2 kolumny z data}

\item{two_wartosc}{double - wyswietlenie emisji}
}
\description{
Funkcja do Wizualizacji danych dostarczonych z pakietem
}
